# Start/stop measurements before and after every type of job