The current execution flow is divided on the following stages:

_start measurements -> job -> stop measurements -> calculate elapsedTime -> index jobSummary ->  evaluate alerts -> indexMetrics_

meaning that elapsedTime was taking into account the "stop measurements" stage

This change proposed to stop Measurements after elapsedTime is calculated, so we won't take into account that time when indexing the job summary document

Fixes: https://github.com/cloud-bulldozer/kube-burner/issues/312