@afcollins 2023-05-24
Maybe it's an error that was masked previously, but there is a measurement error now being logged due to being over the threshold. If I increase the threshold, the error goes away as expected.

master
```
time="2023-05-24 16:19:32" level=info msg="Stopping measurement: podLatency" file="factory.go:109"
time="2023-05-24 16:19:32" level=info msg="Evaluating latency thresholds" file="metrics.go:57"
time="2023-05-24 16:19:32" level=error msg="❗ P99 Ready latency (11.63s) higher than configured threshold: 5s" file="metrics.go:67"
time="2023-05-24 16:19:32" level=info msg="node-density: ContainersReady 50th: 9737 99th: 11632 max: 11632 avg: 9507" file="pod_latency.go:170"
time="2023-05-24 16:19:32" level=info msg="node-density: Ready 50th: 9737 99th: 11632 max: 11632 avg: 9507" file="pod_latency.go:170"
time="2023-05-24 16:19:32" level=info msg="Job node-density took 13s" file="job.go:163"
```

pr/319
```
time="2023-05-24 16:27:13" level=info msg="Stopping measurement: podLatency" file="factory.go:111"
time="2023-05-24 16:27:13" level=info msg="Evaluating latency thresholds" file="metrics.go:57"
time="2023-05-24 16:27:13" level=error msg="❗ P99 Ready latency (11.13s) higher than configured threshold: 5s" file="metrics.go:67"
time="2023-05-24 16:27:13" level=info msg="node-density: Ready 50th: 9876 99th: 11126 max: 11126 avg: 9574" file="pod_latency.go:170"
time="2023-05-24 16:27:13" level=error msg="Failed measurements: %v[podLatency]" file="job.go:165"
```